Fixed long-standing problem around Idea and Cluster "index"
I used to manually keep an auto-incrementing index field in the Idea
and Cluster entities.  But that was problematic - potentially buggy in
parallelized situations and a pain to maintain.

Solved by using a random number field in Cluster to pick a random
cluster. And solved in Idea by removing it and sorting by the key.